[version] move FactorioApiVersion in `core.version` package
